feat: 폴더 경로 세분화, 레이아웃 통일
