+++ src/components/HashtagNav/index.ts
@@ -1,0 +1,1 @@
+export { HashTagNav} from "./HashtagNav";
