+++ src/components/Header/styled.ts
@@ -3,26 +3,23 @@ import SearchIcon from "../../assets/icon/search.svg";
 import Logo from "../../assets/icon/logo.svg";
 import NotificationIcon from "../../assets/icon/notification.svg";
 
-export const Root = styled.div`
+export const Root = styled.header`
   padding-top: 32px;
   display: flex;
   align-items: center;
 `;
 
 export const SearchInput = styled.input`
-  width: 40vw;
+  width: 681px;
   height: 48px;
   padding-left: 60px;
-  border-radius: 10px;
+  border-radius: 15px;
   border: 0.5px solid #e2e8f0;
   background-image: url(${SearchIcon});
   background-repeat: no-repeat;
   background-position: 20px 50%;
   box-shadow: 0px 2px 8px 0px #f1f5f9;
-  font-size: 17px;
-  &:focus {
-    outline: none;
-  }
+  font-size: 16px;
 `;
 
 export const LogoSection = styled.div`
@@ -47,6 +44,7 @@ export const LogoImg = styled.img.attrs({
 export const LogoTypo = styled.div`
   color: #7c3aed;
   text-align: right;
+  font-family: "Inter";
   font-size: 20px;
   font-weight: 600;
   letter-spacing: -1px;
