+++ src/components/List/ArticleList/index.ts
@@ -1,0 +1,1 @@
+export { Articles } from "./Articles"
