+++ src/components/List/ArticlesTable/index.ts
@@ -1,0 +1,1 @@
+export {ArticlesTable} from "./ArticlesTable"
