+++ src/components/Nav/styled.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-export const Container = styled.div`
+export const Container = styled.nav`
   display: flex;
   flex-direction: column;
   width: 185px;
