+++ src/components/RightWrapper/index.ts
@@ -1,0 +1,1 @@
+export { RightWrapper } from "./Wrapper"
