+++ src/components/RightWrapper/styled.ts
