--- src/components/Wrapper/index.ts
@@ -1,1 +1,0 @@
-export { Wrapper } from "./Wrapper"
