--- src/components/Wrapper2.tsx
@@ -1,6 +1,0 @@
-export const Wrapper2 = () => {
-    return(
-        <>
-        </>
-    )
-}
