+++ src/pages/List/List.tsx
@@ -1,0 +1,13 @@
+import { Articles } from "../../components/List/ArticleList/Articles"
+import { ListContainer } from "./styled";
+
+export const List = () => {
+    return (
+      <>
+        <ListContainer>
+          <Articles />
+        </ListContainer>
+      </>
+    );
+  };
+
