+++ src/pages/List/index.ts
@@ -1,0 +1,1 @@
+export { List } from "./List";
