+++ src/pages/List/styled.ts
@@ -1,0 +1,28 @@
+import styled from "styled-components";
+
+export const TopContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const Body = styled.div`
+  display: flex;
+  flex-direction: row;
+  margin-top: 60px;
+`;
+
+export const LWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  width: 27vw;
+`;
+
+export const ListContainer = styled.div`
+  margin-right: 180px;
+`;
+
+export const RWrapper = styled.div`
+  width: 26vw;
+`;
