+++ src/pages/WriteQuestion/styled.ts
@@ -9,7 +9,7 @@ export const QuestionForm = styled.form`
   flex-direction: column;
   width: 683px;
   height: 800px;
-  margin-right: 240px;
+  margin-right: 180px;
 `;
 
 export const QuestionTitleSection = styled.div`
