+++ src/routes/Layout/Layout.tsx
@@ -1,15 +1,21 @@
 import { Outlet } from "react-router-dom";
+import { HashTagNav } from "../../components/HashtagNav";
 import { Header } from "../../components/Header";
 import { Nav } from "../../components/Nav";
-import { Container, Root } from "./styled";
+import { RightWrapper } from "../../components/RightWrapper";
+import { Container, Root, LeftWrapper } from "./styled";
 
 export const Layout = () => {
   return (
     <Root>
       <Header />
       <Container>
-        <Nav />
+        <LeftWrapper>
+          <Nav />
+          <HashTagNav />
+        </LeftWrapper>
         <Outlet />
+        <RightWrapper/>
       </Container>
     </Root>
   );
